+++ src/cosimulation/simulator_interface.c
@@ -82,7 +82,7 @@ void sw_control_req_callback(PLI_CB_ARG)
      */
     for (size_t i = 0; i < strlen(test_name_binary); i += 8) {
         char letter = 0;
-        for (int j = 0; j < 8; j++)
+        for (size_t j = 0; j < 8; j++)
             if (test_name_binary[i + j] == '1')
                 letter |= 0x1 << (7 - j);
         test_name[i / 8] = letter;
